Changed code to avoid compiler warning in gcc.
--HG--
extra : convert_revision : svn%3Afdd8eb12-d10e-0410-9acb-85c331704f74/trunk%403370

+++ dlib/svm/empirical_kernel_map.h
@@ -124,7 +124,7 @@ namespace dlib
 
 
             // now count how many elements of W are non-zero
-            const long num_not_zero = sum(W>eps);
+            const long num_not_zero = static_cast<long>(sum(W>eps));
 
             // Really, this should never happen.  But I'm checking for good measure.
             if (num_not_zero == 0)
